Move swapbox to pages/swap
