+++ src/pages/Swap/Swapbox/Swapbox.tsx
@@ -3,28 +3,28 @@ import { CoWTrade, Currency, CurrencyAmount, JSBI, RoutablePlatform, Trade } fro
 import { useCallback, useEffect, useMemo, useState } from 'react'
 import styled from 'styled-components'
 
-import { ReactComponent as SwapIcon } from '../../assets/images/swap-icon.svg'
-import { AutoColumn } from '../../components/Column'
-import { SwapPoolTabs } from '../../components/NavigationTabs'
-import { ArrowWrapper, SwitchTokensAmountsContainer, Wrapper } from '../../components/Swap/styleds'
-import { useActiveWeb3React } from '../../hooks'
-import { ApprovalState, useApproveCallbackFromTrade } from '../../hooks/useApproveCallback'
-import { useSwapCallback } from '../../hooks/useSwapCallback'
-import { useHigherUSDValue } from '../../hooks/useUSDValue'
-import { useWrapCallback, WrapState, WrapType } from '../../hooks/useWrapCallback'
-import AppBody from '../../pages/AppBody'
-import { useDerivedSwapInfo, useSwapActionHandlers, useSwapState } from '../../state/swap/hooks'
-import { Field } from '../../state/swap/types'
-import { useAdvancedSwapDetails, useIsExpertMode, useUserSlippageTolerance } from '../../state/user/hooks'
-import { computeFiatValuePriceImpact } from '../../utils/computeFiatValuePriceImpact'
-import { maxAmountSpend } from '../../utils/maxAmountSpend'
-import { computeTradePriceBreakdown, warningSeverity } from '../../utils/prices'
-import { CurrencyInputPanel } from '../CurrencyInputPanel'
-import AdvancedSwapDetailsDropdown from '../Swap/AdvancedSwapDetailsDropdown'
-import confirmPriceImpactWithoutFee from '../Swap/confirmPriceImpactWithoutFee'
-import ConfirmSwapModal from '../Swap/ConfirmSwapModal'
-import SwapButtons from '../Swap/SwapButtons'
-import { TradeDetails } from '../Swap/TradeDetails'
+import { ReactComponent as SwapIcon } from '../../../assets/images/swap-icon.svg'
+import { AutoColumn } from '../../../components/Column'
+import { CurrencyInputPanel } from '../../../components/CurrencyInputPanel'
+import { SwapPoolTabs } from '../../../components/NavigationTabs'
+import AdvancedSwapDetailsDropdown from '../../../components/Swap/AdvancedSwapDetailsDropdown'
+import confirmPriceImpactWithoutFee from '../../../components/Swap/confirmPriceImpactWithoutFee'
+import ConfirmSwapModal from '../../../components/Swap/ConfirmSwapModal'
+import { ArrowWrapper, SwitchTokensAmountsContainer, Wrapper } from '../../../components/Swap/styleds'
+import SwapButtons from '../../../components/Swap/SwapButtons'
+import { TradeDetails } from '../../../components/Swap/TradeDetails'
+import { useActiveWeb3React } from '../../../hooks'
+import { ApprovalState, useApproveCallbackFromTrade } from '../../../hooks/useApproveCallback'
+import { useSwapCallback } from '../../../hooks/useSwapCallback'
+import { useHigherUSDValue } from '../../../hooks/useUSDValue'
+import { useWrapCallback, WrapState, WrapType } from '../../../hooks/useWrapCallback'
+import { useDerivedSwapInfo, useSwapActionHandlers, useSwapState } from '../../../state/swap/hooks'
+import { Field } from '../../../state/swap/types'
+import { useAdvancedSwapDetails, useIsExpertMode, useUserSlippageTolerance } from '../../../state/user/hooks'
+import { computeFiatValuePriceImpact } from '../../../utils/computeFiatValuePriceImpact'
+import { maxAmountSpend } from '../../../utils/maxAmountSpend'
+import { computeTradePriceBreakdown, warningSeverity } from '../../../utils/prices'
+import AppBody from '../../AppBody'
 
 export enum CoWTradeState {
   UNKNOWN, // default
